requirements.txt
langchain
langgraph
groq
python-dotenv
streamlit
pydantic
beautifulsoup4
requests
python-jose
passlib
python-multipart
sqlalchemy
pytest
PyPDF2 
langchain_groq
langchain_community